server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
PyYAML>=6.0
openai>=1.12.0
anthropic>=0.18.0
